img: Allow GAP values of 0 to be specified. Auto-GAP is now special value -1. Allow GAP4A to be specified in IMG.CFG

+++ inc/floppy.h
@@ -84,7 +84,7 @@ struct raw_trk {
     uint16_t sec_off;
     uint16_t data_rate;
     uint16_t rpm;
-    uint8_t gap_2, gap_3, gap_4a;
+    int16_t gap_2, gap_3, gap_4a;
     uint8_t interleave, cskew, hskew;
     uint8_t has_iam:1, is_fm:1, invert_data:1;
 };
